:recycle:Refactor: 코드 리뷰 기반 수정사항 반영
