+++ src/main/java/com/likelion13/artium/domain/auth/service/AuthServiceImpl.java
@@ -37,7 +37,7 @@ public class AuthServiceImpl implements AuthService {
 
   @Override
   @Transactional
-  public TokenResponse login(HttpServletResponse response, LoginRequest loginRequest) {
+  public TokenResponse login(LoginRequest loginRequest) {
 
     User user =
         userRepository
@@ -79,7 +79,7 @@ public class AuthServiceImpl implements AuthService {
       long expiration = jwtProvider.getExpirationTime(accessToken);
       redisTemplate
           .opsForValue()
-          .set("BL:" + accessToken, "logout", expiration, TimeUnit.MILLISECONDS);
+          .set("BL:" + accessToken, "logout", expiration, TimeUnit.SECONDS);
 
       log.info("사용자 '{}'의 액세스 토큰이 만료까지 블랙리스트 처리되었습니다.", username);
