+++ src/main/java/com/likelion13/artium/domain/user/repository/UserRepository.java
@@ -8,7 +8,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.likelion13.artium.domain.user.entity.User;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByUsername(String username);
