+++ src/main/java/com/likelion13/artium/domain/user/service/UserService.java
@@ -25,84 +25,31 @@ import com.likelion13.artium.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class UserService {
+/**
+ * 사용자 관련 주요 기능을 제공하는 서비스 인터페이스입니다.
+ */
+public interface UserService {
 
-  private final UserRepository userRepository;
-  private final PasswordEncoder passwordEncoder;
-  private final UserMapper userMapper;
+  /**
+   * 회원가입 요청을 처리합니다.
+   * <p>
+   * 회원가입 시 중복된 아이디가 존재하면 예외를 발생시키며,
+   * 비밀번호는 인코딩되어 저장됩니다.
+   *
+   * @param request 회원가입 요청 데이터 (아이디, 비밀번호, 닉네임 등)
+   * @return 회원가입 성공 시 생성된 사용자 정보를 담은 응답 DTO
+   * @throws CustomException 중복된 아이디 등 가입 불가 사유 발생 시
+   */
+  SignUpResponse signUp(SignUpRequest request);
 
-  @Transactional
-  public SignUpResponse signUp(SignUpRequest request) {
-    if (userRepository.existsByUsername(request.getUsername())) {
-      throw new CustomException(UserErrorCode.USERNAME_ALREADY_EXISTS);
-    }
-
-    // 비밀번호 인코딩
-    String encodedPassword = passwordEncoder.encode(request.getPassword());
-
-    // 유저 엔티티 생성
-    User user =
-        User.builder()
-            .username(request.getUsername())
-            .password(encodedPassword)
-            .nickname(request.getNickname())
-            .provider("custom")
-            .role(Role.USER)
-            .build();
-
-    User savedUser = userRepository.save(user);
-    log.info("새로운 사용자 생성: {}", savedUser.getUsername());
-
-    return userMapper.toSignUpResponse(savedUser);
-  }
-
-  public User getCurrentUser() {
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-    if (authentication == null || !authentication.isAuthenticated()) {
-      log.error("인증 실패 - 인증 정보 없음");
-      throw new CustomException(UserErrorCode.UNAUTHORIZED);
-    }
-
-    Object principal = authentication.getPrincipal();
-    String username = "";
-
-    try {
-      if (principal instanceof OAuth2User oauthUser) {
-        Object email = oauthUser.getAttribute("email");
-        if (email != null) {
-          username = (String) email;
-        } else {
-          Map<String, Object> kakaoAccount = oauthUser.getAttribute("kakao_account");
-          if (kakaoAccount != null && kakaoAccount.containsKey("email")) {
-            username = (String) kakaoAccount.get("email");
-          }
-        }
-      } else if (principal instanceof String str) {
-        username = str;
-      } else if (principal instanceof UserDetails userDetails) {
-        username = userDetails.getUsername();
-      } else {
-        log.error("인증 실패 - Principal 타입 알 수 없음: {}", principal.getClass());
-        throw new CustomException(UserErrorCode.UNAUTHORIZED);
-      }
-    } catch (Exception e) {
-      log.error("인증 정보 추출 중 오류", e);
-      throw new CustomException(UserErrorCode.UNAUTHORIZED);
-    }
-
-    if (username == null || username.isBlank()) {
-      log.error("인증 실패 - 추출된 username이 null 또는 빈 문자열");
-      throw new CustomException(UserErrorCode.UNAUTHORIZED);
-    }
-
-    log.debug("JWT에서 추출한 email: {}", username);
-
-    return userRepository
-        .findByUsername(username)
-        .orElseThrow(() -> new CustomException(UserErrorCode.USER_NOT_FOUND));
-  }
+  /**
+   * 현재 인증된 사용자의 정보를 조회합니다.
+   * <p>
+   * 인증 정보가 없거나 유효하지 않으면 예외를 던지며,
+   * OAuth2 로그인과 자체 로그인 모두 지원합니다.
+   *
+   * @return 현재 인증된 {@link User} 엔티티
+   * @throws CustomException 인증 실패, 사용자 미존재 시 발생
+   */
+  User getCurrentUser();
 }
